Feature: Image DB in progress

+++ app/api/article/route.ts
@@ -1,5 +1,6 @@
 // @ts-nocheck
 import Article from "@/class/article-class";
+import Image from "@/class/image-class"
 import Github from "@/class/github-class";
 
 import connection from "@/utils/connection";
@@ -50,16 +51,32 @@ export async function POST(req: Request) {
   };
   article.image = "/images/" + article.image.name;
 
+  const image_exist = await Image.exist(article.image);
 
-  const github_manager = new Github();
-  const pushed_pass = await github_manager.push_gihtub_files([image]);
+  if ( !image_exist ) {
+     const new_image = await Image.new({ path: article.image, id_used: [] })
 
-  if ( pushed_pass ) {
-    const new_article = await Article.new(article);
-    return httpResponse(new_article);
+    if ( new_image !== StatusCode.Success ) return httpResponse(new_image)
+
+    const github_manager = new Github();
+    const pushed_pass = await github_manager.push_gihtub_files([image]);
+
+    if ( !pushed_pass ) return httpResponse(StatusCode.InternalError);
   }
 
-  return httpResponse(StatusCode.InternalError);
+  const new_article = await Article.new(article);
+
+  if ( new_article.status !== StatusCode.Success || image_exist ) return httpResponse(new_article.status)
+
+  const image_db = await Image.get(article.image);
+
+  if ( !image_db ) return httpResponse(StatusCode.NotFound) 
+
+  image_db.id_used = [...image_db.id_used, new_article._id]
+
+  const update_image = await Image.update(image_db)
+
+  return httpResponse(new_article.status);
 }
 
 
@@ -87,22 +104,36 @@ export async function PATCH(req: Request) {
     };
     update_article.image = "/images/" + update_article.image.name;
 
-    let image_name = article.image.split("/")[2];
+    const delete_if_needed = await Image.delete_id_needed(article.image, update_article._id);
 
-    const github_manager = new Github();
-    const delete_file = await github_manager.delete_github_file(image_name)
+    if ( delete_if_needed ) {
+      let image_name = article.image.split("/")[2];
 
-    if ( !delete_file ) {
-      console.log("GITHUB ERROR: Error deleting file")
-      return httpResponse(StatusCode.InternalError);
+      const github_manager = new Github();
+      const delete_file = await github_manager.delete_github_file(image_name)
+
+      if ( !delete_file ) {
+        console.log("GITHUB ERROR: Error deleting file")
+        return httpResponse(StatusCode.InternalError);
+      }
     }
 
-    const pushed_pass = await github_manager.push_gihtub_files([image]);
 
-    if ( !pushed_pass ) {
-      console.log("GITHUB ERROR: Error pushing file")
-      return httpResponse(StatusCode.InternalError);
+    const image_exist = await Image.exist(update_article.image);
+
+    if ( !image_exist) {
+      const new_image = await Image.new({ path: update_article.image, id_used: [update_article.id] })
+
+      if ( new_image !== StatusCode.Success ) return httpResponse(new_image)
+
+        const pushed_pass = await github_manager.push_gihtub_files([image]);
+
+      if ( !pushed_pass ) {
+        console.log("GITHUB ERROR: Error pushing file")
+        return httpResponse(StatusCode.InternalError);
+      }
     }
+
   } else if ( 
     typeof update_article.image !== "string" && 
     `/images/${update_article.image.name}` === article.image
